refactor(platform): move the Keychain broker out of the platform owner
src/platform is a facade-only owner — everything reachable through
src/platform/index.js is also in the local-review artifact's reviewed graph,
and that graph's source policy forbids any contribution-* module
(scripts/build-local-review-artifact.js, FORBIDDEN_SOURCE_BASENAMES). A broker
living under that owner could only be imported through the barrel, so the
owner boundary and the review boundary were in direct conflict: the barrel
re-export pulled the contribution surface into the reviewed export path, and
importing the module directly tripped the owner rule instead.

Moving it to src/contribution-device-keychain-broker.js, beside the other
contribution modules, satisfies both. This restores three checks the branch
had been failing since the broker landed: the local-review source policy, the
platform adapters' exact public API, and the client exporter's import
allow-list.
